Fix security vulnerability in destroy() method in CommentController

+++ app/Http/Controllers/CommentController.php
@@ -2,13 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\UserRole;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CommentController extends Controller
 {
     public function store(Request $request) 
     {
+        
         $comment = new Comment();
         $comment->fill($request->all());
         $comment->save();
@@ -18,7 +21,10 @@ class CommentController extends Controller
     public function destroy(int $id) 
     {
         $comment = Comment::findOrFail($id);
-        $comment->delete();
-        return redirect()->back();
+        if(Auth::user()->role == UserRole::ADMIN || $comment->user_id == Auth::user()-> id) {
+            $comment->delete();
+            return redirect()->back();
+        }
+        abort(403);
     }
 }
